# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estate Aggregator",
        "description": "An API of app",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/post": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Index",
                "operationId": "b4b7284c8fb22330a82699a8356a5e88",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title post",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Сайт"
                    },
                    {
                        "name": "district_id",
                        "in": "query",
                        "description": "District id",
                        "schema": {
                            "type": "number"
                        },
                        "example": "1702"
                    },
                    {
                        "name": "rubric_id",
                        "in": "query",
                        "description": "Rubric id",
                        "schema": {
                            "type": "number"
                        },
                        "example": "2"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "schema": {
                            "type": "number"
                        },
                        "example": "2"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit data",
                        "schema": {
                            "type": "number"
                        },
                        "example": "20"
                    },
                    {
                        "name": "extends[]",
                        "in": "query",
                        "description": "Extends data",
                        "schema": {
                            "type": "array",
                            "items": {},
                            "example": [
                                "images",
                                "main_image",
                                "user",
                                "district",
                                "rubric"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PostSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Store",
                "operationId": "2b5846c8a1e7a22fa31af93f7accb003",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "district_id",
                                    "rubric_id",
                                    "source"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Сайт"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Хорошо делайте, плохо не делайте"
                                    },
                                    "district_id": {
                                        "type": "number",
                                        "example": "1702"
                                    },
                                    "rubric_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "source": {
                                        "type": "number",
                                        "example": "<a href>сайт</a>"
                                    },
                                    "main_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "images[]": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PostSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The title field is required. (and 2 more errors)"
                                        },
                                        "errors": {
                                            "properties": {
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The title title is required."
                                                    },
                                                    "collectionFormat": "multi"
                                                },
                                                "content": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The content field is required."
                                                    },
                                                    "collectionFormat": "multi"
                                                },
                                                "district_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The district_id field is required."
                                                    },
                                                    "collectionFormat": "multi"
                                                },
                                                "rubric_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The rubric_id field is required."
                                                    },
                                                    "collectionFormat": "multi"
                                                },
                                                "source": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The source field is required."
                                                    },
                                                    "collectionFormat": "multi"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PostSchema": {
                "properties": {
                    "id": {
                        "type": "number",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Почему надо продавать"
                    },
                    "content": {
                        "type": "string",
                        "example": "Да потому что это принесет вам прибыль"
                    },
                    "user_id": {
                        "type": "number",
                        "example": 1
                    },
                    "district_id": {
                        "type": "number",
                        "example": 1702
                    },
                    "rubric_id": {
                        "type": "number",
                        "example": 1
                    },
                    "source": {
                        "type": "string",
                        "example": "<a href=\">источник<a>"
                    },
                    "count_view": {
                        "type": "number",
                        "example": 100
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2022-06-28 06:06:17"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2022-06-28 06:06:17"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}